Change repo_id generation scheme
